Zomato Data Exploration and Analysis with SQL
Most of us know that Zomato is an Indian multinational restaurant aggregator and food delivery company. 
The idea of analysing the Zomato_dataset is to get the overview of what actutally is happening in their business. 
Zomato Dataset consist of more than 9000 rows with columns such as Restaurants_id, Restaurants_name, City, Location, Cuisines and many more...